// src/app/models/producto.module.ts
export interface Producto {
    
    productName: string;
    productType: string;
    description: string;
    energeticValue: number;
    carbohydrates: number;
    protein: number;
    fat: number;
    cholesterol: number;
    sodium: number;
    fiber: number;
    calcium: number;
    vitaminA: number;
    vitaminB1: number;
    vitaminB2: number;
    vitaminB3: number;
    vitaminB5: number;
    vitaminC: number;
    vitaminD: number;
    vitaminE: number;
    vitaminK: number;
    image: string;
}